cleaning

+++ snd/sn76496.c
@@ -11,19 +11,19 @@
 struct SN76496
 {
 	unsigned int UpdateStep;
-	INT32 VolTable[16];	/* volume table         */
-	INT32 Register[8];	/* registers */
-	INT32 LastRegister;	/* last register written */
-	INT32 Volume[4];	/* volume of voice 0-2 and noise */
-	INT32 RNG;		/* noise generator      */
-	INT32 NoiseMode;	/* active noise mode */
-	INT32 FeedbackMask;     /* mask for feedback */
-	INT32 WhitenoiseTaps;   /* mask for white noise taps */
-	INT32 WhitenoiseInvert; /* white noise invert flag */
+	UINT16 VolTable[16];	/* volume table         */
+	UINT16 Register[8];	/* registers */
+	UINT16 LastRegister;	/* last register written */
+	UINT16 Volume[4];	/* volume of voice 0-2 and noise */
+	UINT16 RNG;		/* noise generator      */
+	UINT16 NoiseMode;	/* active noise mode */
+	UINT16 FeedbackMask;     /* mask for feedback */
+	UINT16 WhitenoiseTaps;   /* mask for white noise taps */
+	UINT16 WhitenoiseInvert; /* white noise invert flag */
 	INT32 Period[4];
 	INT32 Count[4];
 	INT32 Output[4];
-	int bSignalAdd;
+	UINT16 bSignalAdd;
 };
 
 struct SN76496 Chip0 = { .LastRegister = 0, .RNG = 0, .NoiseMode = 0, .FeedbackMask = 0, .bSignalAdd = 0, .WhitenoiseInvert = 0, .WhitenoiseTaps = 0 };
@@ -287,7 +287,7 @@ end:
 	{
 		/* limit volume to avoid clipping */
 		if (Out > MAX_OUTPUT / 3) R->VolTable[i] = MAX_OUTPUT / 3;
-		else R->VolTable[i] = (INT32)Out;
+		else R->VolTable[i] = (UINT16)Out;
 
 		Out /= ((double)1.258925412);	/* = 10 ^ (2/20) = 2dB */
 	}
